Edit onPlayerMove click handler
Add functionality of set-div-value.js into onPlayerMove and set a variable
playerToken that can later be switched by another function. Declare variable
data set as an object with two attributes: index, whose value is set to the
number of the id of the clicked div, and value, the text of the clicked div.
This object can now be used in an AJAX PATCH.