Add or edit multiple global functions
Add switchPlayer to toggle between x and o.

Edit testColumns, testRows and testDiagonals to prevent endless loop hangup,
and to prevent checking for wins when an empty string is present at a given
game board index.

Edit whoIsWinner to prevent returning anything other than a string with a value
of x or o.